+++ firmware/src/main.cpp
@@ -73,7 +73,7 @@ LOG_MODULE(main)
  * through the SPI, and then pushed inside the while loop. */
 uint8_t virtual_leds[NUM_PIXELS][3];
 float accel_data[3];
-uint8_t gyro_data[3];
+float gyro_data[3];
 
 /* the pointer inside the LEDFilter class, the base class
  * for LEDFilters which transform motion data into LED patterns,
@@ -81,7 +81,7 @@ uint8_t gyro_data[3];
  * the main program loop, as it is being manipulated inside of the
  * LEDFilter. */
 float *LEDFilter::p_accel_data = accel_data;
-uint8_t *LEDFilter::p_gyro_data = gyro_data;
+float *LEDFilter::p_gyro_data = gyro_data;
 
 uint8_t (*LEDFilter::p_virtual_leds)[3] = virtual_leds;
